refactor func call API về đúng component sử dụng redux, không để tất cả ở component cha, rồi gọi từ component dưới nên nữa.
