+++ src/apis/index.js
@@ -2,10 +2,11 @@ import axios from "axios";
 import{API_ROOT} from "../utils/constants";
 
 //----------------Board API------------------------------
-export const fetchBoardDetailsAPI = async (boardId) => {
-  const response = await axios.get(`${API_ROOT}/v1/boards/${boardId}`);
-  return response.data;
-};
+// đã move vào redux
+// export const fetchBoardDetailsAPI = async (boardId) => {
+//   const response = await axios.get(`${API_ROOT}/v1/boards/${boardId}`);
+//   return response.data;
+// };
 export const updateBoardDetailsAPI = async (boardId, updateData) => {
   const response = await axios.put(`${API_ROOT}/v1/boards/${boardId}`, updateData);
   return response.data;
